+++ src/helpers/Experimental/pipeline/helpers/addDepipe.ts
@@ -1,7 +1,7 @@
-import { depipe } from '../depipe'
 import type { WithDepipe } from '../types/WithDepipe'
 
-/** TODO: create a getDepipeFn to decouple from exported depipe helper */
+import { depipe } from './depipe'
+
 export function addDepipe<RValue>(rvalue: RValue | Promise<RValue>): WithDepipe<typeof rvalue> {
     return Object.defineProperty(Object(rvalue), 'depipe', {
         configurable: true,
